% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BirthWeights.R
\docType{data}
\name{BirthWeights}
\alias{BirthWeights}
\title{BirthWeights data set}
\format{
An object of class \code{"data.frame"}
\describe{
 \item{group}{male or female}
 \item{weights}{BirthWeights for 12 male and  12 female babies}
}
}
\usage{
data(BirthWeights)
}
\description{
Data set on Birthweights for 12 male and 12 female babies as presented by
Dobson (1983, p.14).
}
\examples{
data(BirthWeights)
head(BirthWeights)

}
\references{
Dobson, A.J. (1983). Introduction to Statistical Modelling. Chapman and Hall, London, p.14.
}
\keyword{datasets}
