% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChemicalReaction.R
\docType{data}
\name{ChemicalReaction}
\alias{ChemicalReaction}
\title{Chemical reaction of two methods}
\format{
An object of class \code{"data.frame"}
\describe{
 \item{method}{Method X or Y}
 \item{value}{Chemical reaction values }
}
}
\usage{
data(ChemicalReaction)
}
\description{
This data set is presented by Box et al. (1978, p. 159), where a chemical
reaction was studied by making 10 runs with a standard method X, and 10 runs with
a new, supposedly improved method Y.
}
\examples{
data(ChemicalReaction)
head(ChemicalReaction)

}
\references{
Box, G.E.P, Hunter, W.G. and Hunter, J.S. (1978). Statistics for Experimenters: An Introduction to Design, Data Analysis and Model Building. New York, Wiley, p.159.
}
\keyword{datasets}
