% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FourSources.R
\docType{data}
\name{FourSources}
\alias{FourSources}
\title{Four sources}
\format{
An object of class \code{"data.frame"}
\describe{
 \item{source}{Source A, B, C or D}
 \item{value}{values given for these sources}
}
}
\usage{
data(FourSources)
}
\description{
This data set is used by Coolen and van der Laan (2001)
to introduce NPI for multiple comparisons.
}
\examples{
data(FourSources)
head(FourSources)

}
\references{
Coolen F.P.A. and van der Laan P. (2001). Imprecise predictive selection based on low structure assumptions. Journal of Statistical Planning and Inference, 98(1-2), 259–277.
}
\keyword{datasets}
